refactor: move handlers to dedicated folder
