+++ src/handlers/access.ts
@@ -2,7 +2,11 @@ import { Context } from "telegraf";
 import { Chat, Message } from "telegraf/types";
 import { ConfigChatType } from "../types.ts";
 import { useConfig } from "../config.ts";
-import { getCtxChatMsg, isAdminUser, sendTelegramMessage } from "./telegram.ts";
+import {
+  getCtxChatMsg,
+  isAdminUser,
+  sendTelegramMessage,
+} from "../helpers/telegram.ts";
 import { log } from "../helpers.ts";
 
 export default async function checkAccessLevel(
