+++ src/handlers/onAudio.ts
@@ -4,8 +4,8 @@ import fs from "fs";
 import tmp from "tmp";
 import onTextMessage from "./onTextMessage.ts";
 import checkAccessLevel from "./access.ts";
-import { sendTelegramMessage } from "./telegram.ts";
-import { convertToMp3, sendAudioWhisper } from "./stt.ts";
+import { sendTelegramMessage } from "../helpers/telegram.ts";
+import { convertToMp3, sendAudioWhisper } from "../helpers/stt.ts";
 import { useConfig } from "../config.ts";
 import { log } from "../helpers.ts";
 import { Message } from "telegraf/types";
